refactor(eslint-config): streamline revamped Markdown support
- Remove obsolete processor options from Markdown configuration
- Simplify code block handling by enabling default extraction
- Update glob patterns for better file matching
- Refactor tests to align with new configuration structure
- Remove unused markdown code block configuration file
