+++ packages/eslint-config/src/config.d.ts
@@ -44,9 +44,10 @@ export type ConfigNames =
   | '@bfra.me/jsonc/json-schema'
   | '@bfra.me/jsonc'
   | '@bfra.me/markdown/plugin'
-  | '@bfra.me/markdown/language'
-  | '@bfra.me/markdown/rules'
+  | '@bfra.me/markdown/processor'
+  | '@bfra.me/markdown/parser'
   | '@bfra.me/markdown/disabled'
+  | '@bfra.me/markdown/overrides'
   | '@bfra.me/markdown/code-blocks/typescript-javascript'
   | '@bfra.me/markdown/code-blocks/json'
   | '@bfra.me/markdown/code-blocks/yaml'
