+++ packages/eslint-config/src/configs/index.ts
@@ -9,7 +9,6 @@ export * from './javascript'
 export * from './jsdoc'
 export * from './jsonc'
 export * from './markdown'
-export * from './markdown-code-blocks'
 export * from './nextjs'
 export * from './node'
 export * from './package-json'
