+++ packages/eslint-config/src/configs/markdown.ts
@@ -2,11 +2,21 @@ import type {Plugin} from '@eslint/core'
 import type {Config} from '../config'
 import type {Flatten, OptionsFiles, OptionsOverrides} from '../options'
 import {mergeProcessors, processorPassThrough} from 'eslint-merge-processors'
-import {GLOB_MARKDOWN, GLOB_MARKDOWN_IN_MARKDOWN} from '../globs'
-import {requireOf} from '../require-of'
+import {
+  GLOB_JS,
+  GLOB_JSON,
+  GLOB_JSON5,
+  GLOB_JSONC,
+  GLOB_JSX,
+  GLOB_MARKDOWN,
+  GLOB_MARKDOWN_FILES,
+  GLOB_MARKDOWN_IN_MARKDOWN,
+  GLOB_TS,
+  GLOB_TSX,
+  GLOB_YAML,
+} from '../globs'
+import {plainParser} from '../parsers/plain-parser'
 import {interopDefault} from '../utils'
-import {fallback} from './fallback'
-import {markdownCodeBlocks} from './markdown-code-blocks'
 
 /**
  * Markdown language mode for parsing.
@@ -47,47 +57,6 @@ export type MarkdownLanguage = 'commonmark' | 'gfm'
  * ```
  */
 export type MarkdownFrontmatterOptions = false | 'yaml' | 'toml' | 'json'
-
-/**
- * Processor configuration for Markdown code block extraction and linting.
- *
- * The processor extracts fenced code blocks and lints them as separate virtual files,
- * enabling language-specific rules to apply to code examples.
- *
- * **⚠️ Warning**: Disabled by default due to compatibility issues with ESLint plugins
- * that require full SourceCode API support. Enable with caution and test thoroughly.
- *
- * @example
- * ```typescript
- * // Enable processor with code block extraction
- * defineConfig({
- *   markdown: {
- *     processor: { enabled: true, extractCodeBlocks: true }
- *   }
- * })
- *
- * // Disable code block processing
- * defineConfig({ markdown: { processor: { enabled: false } } })
- * ```
- */
-export interface MarkdownProcessorOptions {
-  /**
-   * Enable the Markdown processor.
-   *
-   * @default false
-   */
-  enabled?: boolean
-
-  /**
-   * Extract and lint code blocks from Markdown files.
-   *
-   * When enabled, fenced code blocks (```ts, ```js, etc.) are extracted
-   * and linted as separate virtual files.
-   *
-   * @default false
-   */
-  extractCodeBlocks?: boolean
-}
 
 /**
  * Code block processing configuration for Markdown files.
@@ -155,7 +124,6 @@ export interface MarkdownCodeBlockOptions {
  *   markdown: {
  *     language: 'gfm',
  *     frontmatter: 'yaml',
- *     processor: { enabled: true, extractCodeBlocks: true },
  *     codeBlocks: { typescript: true, javascript: true }
  *   }
  * })
@@ -194,35 +162,13 @@ export interface MarkdownOptions extends Flatten<OptionsFiles & OptionsOverrides
   frontmatter?: MarkdownFrontmatterOptions
 
   /**
-   * Processor configuration for code block extraction and linting.
-   *
-   * Extracts fenced code blocks and lints them as separate virtual files.
-   *
-   * **⚠️ Warning**: Disabled by default due to compatibility issues with ESLint plugins
-   * requiring full SourceCode API support.
-   *
-   * @default { enabled: false, extractCodeBlocks: false }
-   */
-  processor?: MarkdownProcessorOptions
-
-  /**
    * Code block processing configuration.
    *
    * Configure which languages are extracted and linted from code blocks.
    *
-   * @default { typescript: true, javascript: true, jsx: true }
+   * @default true (enables all supported languages)
    */
-  codeBlocks?: MarkdownCodeBlockOptions
-
-  /**
-   * Markdown-specific rule overrides.
-   *
-   * @example
-   * ```typescript
-   * { rules: { 'markdown/no-html': 'off' } }
-   * ```
-   */
-  rules?: Config['rules']
+  codeBlocks?: boolean | MarkdownCodeBlockOptions
 }
 
 /**
@@ -236,118 +182,225 @@ export interface MarkdownOptions extends Flatten<OptionsFiles & OptionsOverrides
  */
 export async function markdown(options: MarkdownOptions = {}): Promise<Config[]> {
   const {
-    codeBlocks = {javascript: true, jsx: true, typescript: true},
+    codeBlocks = true,
     files = [GLOB_MARKDOWN],
     frontmatter = 'yaml',
     language = 'gfm',
-    processor = {enabled: false, extractCodeBlocks: false},
-    rules = {},
+    overrides = {},
   } = options
 
-  return requireOf(
-    ['@eslint/markdown'],
-    async (): Promise<Config[]> => {
-      const markdown = await interopDefault(import('@eslint/markdown'))
+  const markdown = await interopDefault(import('@eslint/markdown'))
 
-      const configs: Config[] = []
-
-      configs.push({
-        name: '@bfra.me/markdown/plugin',
-        plugins: {
-          markdown: markdown as unknown as Plugin,
-        },
-      })
-      const languageConfig: Config = {
-        name: '@bfra.me/markdown/language',
-        files,
-        ignores: [GLOB_MARKDOWN_IN_MARKDOWN],
-        language: `markdown/${language}`,
-        ...(frontmatter !== false && {
-          languageOptions: {frontmatter},
-        }),
-        ...(processor.enabled !== false && {
-          processor:
-            processor.extractCodeBlocks === false
-              ? processorPassThrough
-              : mergeProcessors([markdown.processors.markdown, processorPassThrough]),
-        }),
-      }
-
-      configs.push(languageConfig)
-
-      configs.push({
-        name: '@bfra.me/markdown/rules',
-        files,
-        rules: {
-          'markdown/fenced-code-language': 'warn',
-          'markdown/heading-increment': 'error',
-          'markdown/no-duplicate-definitions': 'error',
-          'markdown/no-empty-definitions': 'error',
-          'markdown/no-empty-images': 'error',
-          'markdown/no-empty-links': 'error',
-          'markdown/no-invalid-label-refs': 'error',
-          'markdown/no-missing-atx-heading-space': 'error',
-          'markdown/no-missing-label-refs': 'error',
-          'markdown/no-missing-link-fragments': 'error',
-          'markdown/no-multiple-h1': 'error',
-          'markdown/no-reference-like-urls': 'error',
-          'markdown/no-reversed-media-syntax': 'error',
-          'markdown/no-space-in-emphasis': 'error',
-          'markdown/no-unused-definitions': 'error',
-          'markdown/require-alt-text': 'error',
-          'markdown/table-column-count': 'error',
-          ...rules,
-        },
-      })
-
-      // Disable rules incompatible with Markdown processor's virtual files
-      // The @eslint/markdown processor creates virtual files that lack complete ESLint SourceCode API
-      // (getAllComments, getTokenBefore, etc.) and TypeScript parser services (esTreeNodeToTSNodeMap).
-      // This causes failures in plugins that depend on these features.
-      configs.push({
-        name: '@bfra.me/markdown/disabled',
-        files,
-        rules: {
-          'jsdoc/check-param-names': 'off',
-          'jsdoc/check-property-names': 'off',
-          'jsdoc/require-property-name': 'off',
-          'perfectionist/sort-named-exports': 'off',
-          'perfectionist/sort-named-imports': 'off',
-          'unicorn/filename-case': 'off',
-          'command/command': 'off',
-          'jsdoc/check-access': 'off',
-          'jsdoc/check-alignment': 'off',
-          'jsdoc/check-types': 'off',
-          'jsdoc/empty-tags': 'off',
-          'jsdoc/implements-on-classes': 'off',
-          'jsdoc/multiline-blocks': 'off',
-          'jsdoc/no-defaults': 'off',
-          'jsdoc/no-multi-asterisks': 'off',
-          'jsdoc/require-property': 'off',
-          'jsdoc/require-property-description': 'off',
-          'jsdoc/require-returns-check': 'off',
-          'jsdoc/require-returns-description': 'off',
-          'jsdoc/require-yields-check': 'off',
-          'no-irregular-whitespace': 'off',
-          'perfectionist/sort-exports': 'off',
-          'perfectionist/sort-imports': 'off',
-          'regexp/no-legacy-features': 'off',
-          'regexp/no-missing-g-flag': 'off',
-          'regexp/no-super-linear-backtracking': 'off',
-          'regexp/no-useless-dollar-replacements': 'off',
-          'regexp/no-useless-flag': 'off',
-          'regexp/optimal-quantifier-concatenation': 'off',
-          'regexp/require-unicode-sets-regexp': 'off',
-        },
-      })
-
-      // Delegate to specialized module for language-specific code block handling
-      if (processor.extractCodeBlocks !== false) {
-        configs.push(...(await markdownCodeBlocks(codeBlocks)))
-      }
-
-      return configs
+  const configs: Config[] = [
+    {
+      name: '@bfra.me/markdown/plugin',
+      plugins: {
+        markdown: markdown as unknown as Plugin,
+      },
     },
-    fallback,
-  )
+    {
+      name: '@bfra.me/markdown/processor',
+      files,
+      ignores: [GLOB_MARKDOWN_IN_MARKDOWN],
+      language: `markdown/${language}`,
+      processor: mergeProcessors([markdown.processors.markdown, processorPassThrough]),
+      ...(frontmatter !== false && {
+        languageOptions: {frontmatter},
+      }),
+    },
+    {
+      name: '@bfra.me/markdown/parser',
+      files,
+      languageOptions: {
+        parser: plainParser,
+      },
+    },
+
+    // Disable rules incompatible with Markdown processor's virtual files
+    // The @eslint/markdown processor creates virtual files that lack complete ESLint SourceCode API
+    // (getAllComments, getTokenBefore, etc.) and TypeScript parser services (esTreeNodeToTSNodeMap).
+    // This causes failures in plugins that depend on these features.
+    {
+      name: '@bfra.me/markdown/disabled',
+      files,
+      rules: {
+        'jsdoc/check-property-names': 'off',
+        'jsdoc/require-property-name': 'off',
+
+        'unicorn/filename-case': 'off',
+
+        'command/command': 'off',
+
+        'jsdoc/check-access': 'off',
+        'jsdoc/check-types': 'off',
+        'jsdoc/empty-tags': 'off',
+        'jsdoc/no-multi-asterisks': 'off',
+        'jsdoc/require-property': 'off',
+        'jsdoc/require-property-description': 'off',
+
+        'no-irregular-whitespace': 'off',
+
+        'perfectionist/sort-exports': 'off',
+        'perfectionist/sort-imports': 'off',
+
+        'regexp/no-legacy-features': 'off',
+        'regexp/no-missing-g-flag': 'off',
+        'regexp/no-useless-dollar-replacements': 'off',
+        'regexp/no-useless-flag': 'off',
+      },
+    },
+    {
+      name: '@bfra.me/markdown/overrides',
+      files,
+      rules: {
+        'markdown/fenced-code-language': 'warn',
+        'markdown/heading-increment': 'error',
+        'markdown/no-duplicate-definitions': 'error',
+        'markdown/no-empty-definitions': 'error',
+        'markdown/no-empty-images': 'error',
+        'markdown/no-empty-links': 'error',
+        'markdown/no-invalid-label-refs': 'error',
+        'markdown/no-missing-atx-heading-space': 'error',
+        'markdown/no-missing-label-refs': 'error',
+        'markdown/no-missing-link-fragments': 'error',
+        'markdown/no-multiple-h1': 'error',
+        'markdown/no-reference-like-urls': 'error',
+        'markdown/no-reversed-media-syntax': 'error',
+        'markdown/no-space-in-emphasis': 'error',
+        'markdown/no-unused-definitions': 'error',
+        'markdown/require-alt-text': 'error',
+        'markdown/table-column-count': 'error',
+
+        ...overrides,
+      },
+    },
+  ]
+
+  if (codeBlocks === false) {
+    return configs
+  }
+
+  const {
+    javascript = true,
+    json = true,
+    jsx = true,
+    typescript = true,
+    yaml = true,
+  } = codeBlocks === true ? {} : codeBlocks
+
+  if (typescript || javascript || jsx) {
+    const tselint = await interopDefault(import('typescript-eslint'))
+
+    const tsJsFiles: string[] = []
+    if (typescript) {
+      tsJsFiles.push(...GLOB_MARKDOWN_FILES.map(p => `${p}/${GLOB_TS}`))
+    }
+    if (typescript || jsx) {
+      tsJsFiles.push(...GLOB_MARKDOWN_FILES.map(p => `${p}/${GLOB_TSX}`))
+    }
+    if (javascript) {
+      tsJsFiles.push(...GLOB_MARKDOWN_FILES.map(p => `${p}/${GLOB_JS}`))
+    }
+    if (javascript || jsx) {
+      tsJsFiles.push(...GLOB_MARKDOWN_FILES.map(p => `${p}/${GLOB_JSX}`))
+    }
+
+    configs.push({
+      name: '@bfra.me/markdown/code-blocks/typescript-javascript',
+      files: tsJsFiles,
+      languageOptions: {
+        parser: tselint.parser,
+        parserOptions: {
+          ecmaFeatures: {
+            impliedStrict: true,
+            jsx: jsx || javascript,
+          },
+          ecmaVersion: 'latest',
+          // Type-aware rules disabled: documentation snippets lack tsconfig context
+          project: null,
+          sourceType: 'module',
+        },
+      },
+      rules: {
+        // Only disable non-type-aware rules we want to skip for markdown code blocks
+        '@typescript-eslint/no-namespace': 'off',
+
+        '@stylistic/comma-dangle': 'off',
+        '@stylistic/eol-last': 'off',
+        '@stylistic/padding-line-between-statements': 'off',
+
+        '@typescript-eslint/consistent-type-imports': 'off',
+        '@typescript-eslint/explicit-function-return-type': 'off',
+        '@typescript-eslint/no-redeclare': 'off',
+        '@typescript-eslint/no-require-imports': 'off',
+        '@typescript-eslint/no-unused-expressions': 'off',
+        '@typescript-eslint/no-unused-vars': 'off',
+        '@typescript-eslint/no-use-before-define': 'off',
+        '@typescript-eslint/no-var-requires': 'off',
+
+        'import-x/newline-after-import': 'off',
+
+        'jsdoc/require-returns-check': 'off',
+
+        'no-alert': 'off',
+        'no-console': 'off',
+        'no-labels': 'off',
+        'no-lone-blocks': 'off',
+        'no-restricted-imports': 'off',
+        'no-restricted-syntax': 'off',
+        'no-undef': 'off',
+        'no-unused-expressions': 'off',
+        'no-unused-labels': 'off',
+        'no-unused-vars': 'off',
+
+        'node/prefer-global/process': 'off',
+
+        'unicode-bom': 'off',
+
+        'unused-imports/no-unused-imports': 'off',
+        'unused-imports/no-unused-vars': 'off',
+      },
+    })
+  }
+
+  if (json) {
+    const pluginJsonc = await interopDefault(import('eslint-plugin-jsonc'))
+
+    // Use plugin's base config to inherit parser configuration
+    const baseConfigs = pluginJsonc.configs['flat/base']
+
+    configs.push({
+      name: '@bfra.me/markdown/code-blocks/json',
+      files: GLOB_MARKDOWN_FILES.flatMap(p => [
+        `${p}/${GLOB_JSON}`,
+        `${p}/${GLOB_JSON5}`,
+        `${p}/${GLOB_JSONC}`,
+      ]),
+      languageOptions: baseConfigs[1]?.languageOptions ?? {},
+      rules: {
+        // Examples may show JSON with trailing commas or explanatory comments
+        'jsonc/comma-dangle': 'off',
+        'jsonc/no-comments': 'off',
+      },
+    })
+  }
+
+  if (yaml) {
+    const pluginYaml = await interopDefault(import('eslint-plugin-yml'))
+
+    // Use plugin's standard config to inherit parser configuration
+    const standardConfigs = pluginYaml.configs['flat/standard'] as Config[]
+
+    configs.push({
+      name: '@bfra.me/markdown/code-blocks/yaml',
+      files: GLOB_MARKDOWN_FILES.map(p => `${p}/${GLOB_YAML}`),
+      languageOptions: standardConfigs[1]?.languageOptions ?? {},
+      rules: {
+        // Examples may show incomplete YAML mappings to focus on specific concepts
+        'yml/no-empty-mapping-value': 'off',
+      },
+    })
+  }
+
+  return configs
 }
