+++ packages/eslint-config/src/globs.ts
@@ -1,10 +1,11 @@
 export const GLOB_SRC_EXT = '?([cm])[jt]s?(x)'
 export const GLOB_SRC = '**/*.?([cm])[jt]s?(x)'
 
+export const GLOB_JS = '**/*.?([cm])js'
 export const GLOB_JSX = '**/*.?([cm])jsx'
 
 export const GLOB_TS = '**/*.?([cm])ts'
-export const GLOB_TSX = '**/*.tsx'
+export const GLOB_TSX = '**/*.?([cm])tsx'
 
 export const GLOB_JSON = '**/*.json'
 export const GLOB_JSON5 = '**/*.json5'
